+++ app/ShippingManifesto.php
@@ -3,15 +3,9 @@
 namespace KushyApi;
 
 use Illuminate\Database\Eloquent\Model;
-use KushyApi\Traits\Uuids;
 
 class ShippingManifesto extends Model
 {
-    /**
-     * Generates and inserts uuid when creating new items
-     */
-    use Uuids;
-
     /**
      * Indicates if the IDs are auto-incrementing.
      *
@@ -66,9 +60,72 @@ class ShippingManifesto extends Model
         'date_actual_arrival',
         'date_signed',
         'digital_signature',
-
-        'access_token',
     ];
+
+    /**
+     * Validation rules
+     *
+     * @var array
+     */
+    protected $rules = [
+        'shipper_id' => 'string|required',
+        'shipper_state_license' => 'string|required',
+        'shipper_name' => 'string|required',
+        'shipper_address' => 'string|required',
+        'shipper_city' => 'string|required',
+        'shipper_state' => 'string|required',
+        'shipper_postal_code' => 'string|required',
+        'shipper_phone' => 'numeric|required',
+
+        'receiver_id' => 'string|required|exists:posts,id',
+        'receiver_state_license' => 'string|required',
+        'receiver_name' => 'string|required',
+        'receiver_address' => 'string|required',
+        'receiver_city' => 'string|required',
+        'receiver_state' => 'string|required',
+        'receiver_postal_code' => 'string|required',
+        'receiver_phone' => 'numeric|required',
+        'receiver_email' => 'email|required',
+
+        'distributor_id' => 'string|required|exists:posts,id',
+        'distributor_state_license' => 'string|required',
+        'distributor_name' => 'string|required',
+        'distributor_address' => 'string|required',
+        'distributor_city' => 'string|required',
+        'distributor_state' => 'string|required',
+        'distributor_postal_code' => 'string|required',
+        'distributor_phone' => 'numeric|required',
+        'distributor_email' => 'email|required',
+
+        // @todo: when delivery drivers are implemented, uncomment below and delete dupe param
+        // 'delivery_id' => 'string|required|exists:delivery_drivers,id',
+        'delivery_id' => 'string|required',
+        'owner_id' => 'string|required',
+        'invoice' => 'string|required',
+        'status' => 'string|required',
+        'description' => 'string|required',
+        'contact_name' => 'string|required',
+        'phone' => 'numeric|required',
+        'date_departure' => 'date|required',
+        'date_estimated_arrival' => 'date|required',
+        'date_actual_arrival' => 'date|required',
+        'date_signed' => 'date|required',
+        'digital_signature' => 'string|required',
+    ];
+
+    /**
+     * Add an access token when creating new manifestos
+     * ++ Generates and inserts uuid when creating new items
+     */
+    protected static function boot()
+    {
+        parent::boot();
+
+        static::creating(function ($model) {
+            $model->{$model->getKeyName()} = (string) \Illuminate\Support\Str::uuid();
+            $model->access_token = md5(date("F j, Y, g:i a"));
+        });
+    }
 
     /**
      * Get the manifesto's products