+++ database/seeds/DeliveryDriversSeeder.php
@@ -1,0 +1,16 @@
+<?php
+
+use Illuminate\Database\Seeder;
+
+class DeliveryDriversSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        factory(KushyApi\DeliveryDrivers::class, 10)->create();
+    }
+}